requirements.txt
praw==7.7.1
python-dotenv
psycopg2-binary